refactor(LottoMachine): 랜덤 로또를 생성하는 메서드를 LottoGenerator 클래스로 분리
