--- src/constants.js
@@ -1,9 +1,0 @@
-const LOTTO = Object.freeze({
-  RANGE_MIN: 1,
-  RANGE_MAX: 45,
-  COUNT: 6,
-});
-
-module.exports = {
-  LOTTO,
-};
